Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

There is another way where we create a common overlay. I am going to commit this in a little bit such that it works for both I9100 and I777. It involves creating a TARGET_SPECIFIC_HEADER_PATH -- not ideal either because it forks gps.h for our device.

Patch-set: 1