Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I don't think forking gps.h is likely to be too bad - it seems like it changes fairly rarely.

I guess the question is - is there a good way to document this such that people working on other devices who encounter similar problems are aware of this potential change?  (It'll be hard to find if buried in a select few device trees.)  I suspect we'll be seeing this more often now that Russia is putting a 25% import tax on any device that supports GPS and not GLONASS.  (Assuming the extra field is related to multi-constellation support - I am guessing that it is as there's an extra mGnssUsed field in GpsLocationProvider of Samsung firmwares that I saw when baksmali'ing them.)

Patch-set: 1