Update patch set 2
Patch Set 2:

Yes, for 4.1 system/core and hardware/libhardware_legacy needs to be patched. But I will change the patches to just 4.2 support since that is very simple. And I'm hoping that we do not need to fallback to our 4.1 blobs. We can add support for 4.1 later if necessary.

Patch-set: 2