Update patch set 3
Patch Set 3:

The light_state_t structure is the only data gateway between the Java / jni settings and the C LibLights modules, it 's mean to carry variables across the layers. It's compatible with everyone because the other LibLights won't use the variables, unless other devices find it useful and implement the feature. Brightness for example might be interesting for other phones, and multipleLEDs for the SONY Xperias with Illumination Bar (S/P/U/SP).
The other way is to add parameters to each layer of the notification lights and that's non sense because all LibLights have to be recoded, and that why the structure was designed for.

The description of the structure says : "Not all lights must support all parameters."

Patch-set: 3