Update patch set 7
Patch Set 7:

Thanks.

For LIGHT_BRIGHTNESS_MAXIMUM I prefered to put it there to avoid each LibLights using it setting the constant (cleaner that way), but if you want it can go.

The LIGHT_MODE_MULTIPLE_LEDS has to stay, it represents the modes masks and may one day be completed with other masks in 0x02, 0x04, ... and needs to remain the same for every LibLights using it.

Patch-set: 7