Update patch set 1
Patch Set 1:

yes, it resolves the issue. If you don't want to modify the data structure, you might need dynamic allocate temp buffer and free it after usage. The solution is not optimized for frequent memory allocate/free.
Actually you can also fix in the caller side to pass in double size buffer. But personally I think we should fix in the bottom, in case we had multiple callers to modify.

Patch-set: 1