Update patch set 1
Patch Set 1:

In xts google.gmscorehost test, there are some RemoteSubmix related tests which will pop heap corruption error, when the audio is set to mono.

Yes, the reuse of the resampler buffer may not be that appropriate. But the basic idea of the fix is to use one local buffer to store audio from pipe for stereo and do channel conversion and then copy to destination mono buffer. 

Anyway it is not correct to store pipe audio in destination buffer directly. It will overrun the destination buffer.

Patch-set: 1